Crate mithril_client_cli

A command line interface that uses the Mithril Client Library to manipulate Mithril certified types from a Mithril Aggregator:

Cardano DB: List, Show, download and verify

Mithril Stake Distribution: List, download and verify You can find more information on how it works reading the documentation website.

Modules

commands

Command module This module holds the subcommands that can be used from the CLI.

Structs

CommandContext

Context for the command execution

Type Aliases

ClapError

Error Clap