Crate mithril_client_cli

A command line interface that uses the Mithril Client Library to manipulate Mithril certified types from a Mithril Aggregator:

Cardano Database v1 (aka Snapshot): list, show, download archive

Cardano Database v2: list, show, download archive

Cardano transactions: list & show snapshot, certify a list of transactions

Cardano stake distribution: list and download

Mithril stake distribution: list and download

You can find more information on how it works reading the documentation website.

Modules

commands

Command module This module holds the subcommands that can be used from the CLI.

Structs

CommandContext

Context for the command execution

Type Aliases

ClapError

Error Clap