mithril_client_cli
Module commands

Command module This module holds the subcommands that can be used from the CLI.

Modules

cardano_db

Commands for the cardano db artifact

cardano_stake_distribution

Commands for the Cardano Stake Distribution artifact

cardano_transaction

Commands for the Cardano Transaction Snapshot artifact & Cardano Transactions Proof

mithril_stake_distribution

Commands for the Mithril Stake Distribution artifact

Structs

DeprecatedCommand

Stores the deprecated command name and the new command name to use.

Deprecation

Tool to handle deprecated Clap commands.

SharedArgs

Shared arguments for all commands